refactor(api): move user profiles acceptance tests in profile context
